Eliminate mdsvex layouts Fixes #505
