+++ package.json
@@ -37,6 +37,6 @@
 		"sveltekit-embed": "^0.0.5",
 		"tslib": "^2.4.0",
 		"typescript": "^4.7.2",
-		"vite": "^3.1.0-beta.1"
+		"vite": "^3.1.0-beta.2"
 	}
 }
