--- src/lib/elements/h1.svelte
@@ -1,8 +1,0 @@
-<h1><slot /></h1>
-
-<style>
-	h1 {
-		font-size: var(--font-size-fluid-3);
-		line-height: var(--font-lineheight-1);
-	}
-</style>
