--- src/lib/elements/index.js
@@ -1,7 +1,0 @@
-import h1 from './h1.svelte';
-import p from './p.svelte';
-import table from './table.svelte';
-import td from './td.svelte';
-import th from './th.svelte';
-
-export { h1, p, table, td, th };
