--- src/lib/elements/p.svelte
@@ -1,11 +1,0 @@
-<p><slot /></p>
-
-<style>
-	p {
-		margin-block-end: var(--size-fluid-3);
-	}
-
-	p:last-child {
-		margin: 0;
-	}
-</style>
