--- src/lib/elements/td.svelte
@@ -1,12 +1,0 @@
-<td>
-	<slot />
-</td>
-
-<style>
-	td {
-		border: var(--border-size-1) solid var(--text-2);
-		white-space: nowrap;
-		padding-inline: var(--size-3);
-		padding-block: var(--size-2);
-	}
-</style>
