--- src/lib/elements/th.svelte
@@ -1,15 +1,0 @@
-<th>
-	<slot />
-</th>
-
-<style>
-	th {
-		border: var(--border-size-1) solid var(--text-2);
-		background-color: var(--surface-3);
-		font-weight: var(--font-weight-5);
-		text-align: start;
-		white-space: nowrap;
-		padding-inline: var(--size-3);
-		padding-block: var(--size-2);
-	}
-</style>
