+++ src/routes/[topic]/+page.server.ts
@@ -1,5 +1,5 @@
 import topics from '$data/topics';
-import { getPosts } from '$lib/posts';
+import { getPosts } from '$lib/utils/posts';
 import type { PageServerLoad } from './$types';
 
 export const load: PageServerLoad = async function ({ params }) {
