+++ src/routes/posts/[slug=canonical_urls_in_nextjs]/+page.server.ts
