+++ src/routes/posts/[slug=exploring_d3_array]/+page.server.ts
