+++ src/routes/posts/[slug=gatsby_route_api]/+page.server.ts
