+++ src/routes/posts/[slug=lint_staged_config]/+page.server.ts
