+++ src/routes/posts/[slug=sveltekit_and_fathom_analytics]/+page.server.ts
